Feat: Initial proyect commit with: AsyncStorage Api first implementation and some dummy views
Added some views to begin to understand both Tabs and Stack navigators,
we can see the list of decks, add a deck and cards to it and a deck detail.
All this views have dummy data and styling. Next we will focus on adding
better views organized to begin adding afterwards the quiz game.